테스트 발동시? 체인 우라라

+++ HB_Rune/src/main/java/com/ho_bot/HB_Rune/rune/ActiveRune.java
@@ -13,20 +13,20 @@ public class ActiveRune extends Rune{
 	}
 	
 	@Override
-	public void active(Player player ,float amp_val, float amp_per) {
+	public void active(Player player , int amp_val, double amp_per) {
 		if(this.abilityType == AbilityType.Potion) Potion(player, amp_val, amp_per);
 	}
 	
-	private void Potion(Player player, float amp_val, float amp_per) {
+	private void Potion(Player player, int amp_val, double amp_per) {
 		PotionEffectType potion = null;
 		if(this.abilityEffet.equalsIgnoreCase("신속")) potion = PotionEffectType.SPEED;
 		
 		player.addPotionEffect(new PotionEffect(potion, (int) plusAmp(duration, amp_val, amp_per), power));
 	}
 	
-	private <T> float plusAmp(T t, float amp_val, float amp_per) {
+	private <T> double plusAmp(T t, int amp_val, double amp_per) {
 		if(t instanceof Number n) {
-			return (n.floatValue()+amp_val)*(1f+(amp_per/100f));
+			return (n.doubleValue()+amp_val)*(1f+(amp_per/100f));
 		}
 		return 0;
 	}
